Frontend remake.
There are three columns of information right now on the "front page".
We have the left column when there are some demo functions which will be probably removed or replaced in the future, then we have the main camera in center, which will have one camera in a big picture and controls to manage this one stream ( still to do ). At last in the right part of the page we will have the miniatures for all the cameras added to the session, to get one on the front page just click it.